Move GetFeatureFlagsAsync to PostHogClient
